[56] WIP: pytest failing on StorageFactory.get_storage function
This was because get_storage had not been mocked out and provided a return_value yet when the transformer was instantiated.Therefore the transformer init used the original get_storage code, not the mocked return_value provided later.

Fix was to move transformer instantiation to where return_value had been provided.